+++ src/imio/urban/core/contents/parcel/views.py
@@ -1,7 +1,29 @@
 # -*- coding: utf-8 -*-
 
+from plone.dexterity.browser import add
 from plone.dexterity.browser import edit
 from plone.dexterity.browser import view
+
+
+class ParcelAddForm(add.DefaultAddForm):
+    """
+    Parcel custom Add form.
+    """
+
+    portal_type = 'Parcel'
+
+    def __init__(self, context, request):
+        super(ParcelAddForm, self).__init__(context, request)
+        # disable portlets on parcels
+        self.request.set('disable_plone.rightcolumn', 1)
+        self.request.set('disable_plone.leftcolumn', 1)
+
+
+class ParcelAddView(add.DefaultAddView):
+    """
+    Parcel custom Add view.
+    """
+    form = ParcelAddForm
 
 
 class ParcelEditForm(edit.DefaultEditForm):
@@ -11,9 +33,7 @@ class ParcelEditForm(edit.DefaultEditForm):
 
     def __init__(self, context, request):
         super(ParcelEditForm, self).__init__(context, request)
-        self.context = context
-        self.request = request
-        # disable portlets on licences
+        # disable portlets on parcels
         self.request.set('disable_plone.rightcolumn', 1)
         self.request.set('disable_plone.leftcolumn', 1)
 
@@ -25,9 +45,7 @@ class ParcelView(view.DefaultView):
 
     def __init__(self, context, request):
         super(ParcelView, self).__init__(context, request)
-        self.context = context
-        self.request = request
-        # disable portlets on licences
+        # disable portlets on parcels
         self.request.set('disable_plone.rightcolumn', 1)
         self.request.set('disable_plone.leftcolumn', 1)